chore: rm nfts holders and search from route
