+++ packages/react-app-revamp/hooks/useSearchNfts/index.ts
@@ -1,4 +1,29 @@
+import { addressRegex } from "@helpers/regex";
 import { useQuery } from "@tanstack/react-query";
+
+const chainToAlchemySubdomain = {
+  mainnet: "eth-mainnet",
+  polygon: "polygon-mainnet",
+  arbitrumone: "arb-mainnet",
+  optimism: "opt-mainnet",
+  base: "base-mainnet",
+};
+
+const alchemyApiKey = process.env.NEXT_PUBLIC_ALCHEMY_KEY;
+
+const getAlchemyBaseUrl = (chain: string) => {
+  const subdomain = chainToAlchemySubdomain[chain as keyof typeof chainToAlchemySubdomain] || "eth-mainnet";
+  return `https://${subdomain}.g.alchemy.com/nft/v3/${alchemyApiKey}/searchContractMetadata`;
+};
+
+const getAlchemyBaseUrlForContractMetadata = (chain: string) => {
+  const subdomain = chainToAlchemySubdomain[chain as keyof typeof chainToAlchemySubdomain] || "eth-mainnet";
+  return `https://${subdomain}.g.alchemy.com/nft/v3/${alchemyApiKey}/getContractMetadata`;
+};
+
+const NOT_SUPPORTED_NFT_STANDARD = "NO_SUPPORTED_NFT_STANDARD";
+const NOT_A_CONTRACT = "NOT_A_CONTRACT";
+const UNKNOWN = "UNKNOWN";
 
 export interface NFTMetadata {
   address: string;
@@ -10,19 +35,84 @@ export interface NFTMetadata {
   tokenType: string;
 }
 
+function generateSymbolFromName(name: string): string {
+  return name.substring(0, 4).toUpperCase();
+}
+
 const useSearchNfts = (chain: string, query: string) => {
+  const isQueryTokenAddress = addressRegex.test(query);
+
   const fetchNftContractMetadata = async (): Promise<NFTMetadata[]> => {
     if (!query) return [];
 
-    const response = await fetch(
-      `/api/search-nfts?chain=${encodeURIComponent(chain)}&query=${encodeURIComponent(query)}`,
-    );
+    let contracts = [];
+    const baseUrl = getAlchemyBaseUrl(chain);
+    const response = await fetch(`${baseUrl}?query=${encodeURIComponent(query)}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
 
     if (!response.ok) {
-      throw new Error(`network response was not ok: ${response.statusText}`);
+      throw new Error(`Network response was not ok: ${response.statusText}`);
     }
 
-    return response.json();
+    const data = await response.json();
+
+    if (data.contracts && data.contracts.length > 0) {
+      const filteredContracts = data.contracts.filter(
+        (contract: { tokenType: string }) => contract.tokenType !== NOT_SUPPORTED_NFT_STANDARD,
+      );
+
+      contracts = filteredContracts.map((contract: any) => ({
+        address: contract.address,
+        name: contract.openSeaMetadata?.collectionName ? contract.openSeaMetadata.collectionName : contract.name,
+        symbol: contract.symbol ? contract.symbol : generateSymbolFromName(contract.name),
+        totalSupply: contract.totalSupply,
+        tokenType: contract.tokenType,
+        imageUrl: contract.openSeaMetadata?.imageUrl
+          ? contract.openSeaMetadata.imageUrl
+          : "/contest/mona-lisa-moustache.png",
+        isVerified: contract.openSeaMetadata?.safelistRequestStatus === "verified",
+      }));
+    } else if (isQueryTokenAddress) {
+      const contractMetadataUrl = getAlchemyBaseUrlForContractMetadata(chain);
+      const contractResponse = await fetch(`${contractMetadataUrl}?contractAddress=${encodeURIComponent(query)}`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+
+      if (contractResponse.ok) {
+        const contract = await contractResponse.json();
+
+        if (
+          contract.tokenType === NOT_A_CONTRACT ||
+          contract.tokenType === NOT_SUPPORTED_NFT_STANDARD ||
+          contract.tokenType === UNKNOWN
+        ) {
+          return [];
+        }
+
+        contracts = [
+          {
+            address: contract.address,
+            name: contract.openSeaMetadata?.collectionName ? contract.openSeaMetadata.collectionName : contract.name,
+            symbol: contract.symbol ? contract.symbol : generateSymbolFromName(contract.name),
+            totalSupply: contract.totalSupply,
+            tokenType: contract.tokenType,
+            imageUrl: contract.openSeaMetadata?.imageUrl
+              ? contract.openSeaMetadata?.imageUrl
+              : "/contest/mona-lisa-moustache.png",
+            isVerified: contract.openSeaMetadata?.safelistRequestStatus === "verified",
+          },
+        ];
+      }
+    }
+
+    return contracts;
   };
 
   const { data, error, isLoading } = useQuery({
